% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{silo_ppd}
\alias{silo_ppd}
\title{Get PPD data from SILO}
\usage{
silo_ppd(
  station,
  email,
  start = as.Date("1889-01-01"),
  finish = Sys.Date(),
  format = "apsim",
  outfile = NULL
)
}
\arguments{
\item{station}{ppd station id}

\item{email}{email address}

\item{start}{start date}

\item{finish}{finish date}

\item{format}{data format}

\item{outfile}{output file}
}
\value{
No return values
}
\description{
Get PPD data from SILO
}
\details{
Get weather record from SILO PPD (https://www.longpaddock.qld.gov.au/silo/api-documentation/)
}
