% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{silo_ppd}
\alias{silo_ppd}
\title{Retrieve Daily Weather Data from SILO PPD}
\usage{
silo_ppd(
  station,
  email,
  start = as.Date("1889-01-01"),
  finish = Sys.Date(),
  format = c("apsim", "standard"),
  outfile = NULL
)
}
\arguments{
\item{station}{Character. The SILO PPD station ID.}

\item{email}{Character. Email address required for API access.}

\item{start}{Date. The start date for data retrieval.}

\item{finish}{Date. The end date for data retrieval.}

\item{format}{Character. Desired data format (e.g., "csv", "json").}

\item{outfile}{Character. Path to the output file where data will be saved.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effect of downloading and/or saving data.
}
\description{
Downloads daily weather records from the SILO Patched Point Dataset (PPD) using the SILO API.
}
\details{
This function accesses the SILO PPD service (see \url{https://www.longpaddock.qld.gov.au/silo/api-documentation/})
 to retrieve weather data for a specified station and date range. The data can be returned in various formats and
 optionally saved to a file.
}
